Add Delayed Timing
I added delayed timing using 20 sound samples that were 1ms apart from
one another. Using the timing between events, I tried to schedule the
sounds to occur at the correct rate. I'm going to leave it like this
becuase I like the work I've done, but it doesn't actually impact the
timing. Upon further review of the docs, I suspect it is because the
mixer objects playing the sounds run on background processes and are
fairly independent of the event queue.